feat(#138): add package-info.java classes and rename some package
